Refactor iam
